Compartmentalize MainScriptBuilder subclasses into separate files
