+++ SOC/Core/Classes/QuestBuild/Builders/MainScriptBuilder/OnInitialize.cs
@@ -1,0 +1,17 @@
+﻿namespace SOC.Classes.Lua
+{
+    public class OnInitialize
+    {
+        public LuaFunctionBuilder Function = new LuaFunctionBuilder();
+
+        public OnInitialize()
+        {
+            Function.AppendLuaValue(Lua.FunctionCall(Lua.TableIdentifier("TppQuest", "QuestBlockOnInitialize"), Lua.Variable("this")));
+        }
+
+        public LuaTableEntry Get()
+        {
+            return Function.ToTableEntry("OnInitialize", true);
+        }
+    }
+}
