+++ SOC/Core/Classes/QuestBuild/Builders/MainScriptBuilder/OnTerminate.cs
@@ -1,0 +1,17 @@
+﻿namespace SOC.Classes.Lua
+{
+    public class OnTerminate
+    {
+        public LuaFunctionBuilder Function = new LuaFunctionBuilder();
+
+        public OnTerminate()
+        {
+            Function.AppendLuaValue(Lua.FunctionCall(Lua.TableIdentifier("TppQuest", "QuestBlockOnTerminate"), Lua.Variable("this")));
+        }
+
+        public LuaTableEntry Get()
+        {
+            return Function.ToTableEntry("OnTerminate", true);
+        }
+    }
+}
