+++ SOC/Core/Classes/QuestBuild/Builders/MainScriptBuilder/OnUpdate.cs
@@ -1,0 +1,17 @@
+﻿namespace SOC.Classes.Lua
+{
+    public class OnUpdate
+    {
+        public LuaFunctionBuilder Function = new LuaFunctionBuilder();
+
+        public OnUpdate()
+        {
+            Function.AppendLuaValue(Lua.FunctionCall(Lua.TableIdentifier("TppQuest", "QuestBlockOnUpdate"), Lua.Variable("this")));
+        }
+
+        public LuaTableEntry Get()
+        {
+            return Function.ToTableEntry("OnUpdate", true);
+        }
+    }
+}
